// server/src/types/socket.ts
import { Socket } from "socket.io";

// Type definitions
type SocketId = string;

enum SocketEvent {
  JOIN_REQUEST = "join-request",
  JOIN_ACCEPTED = "join-accepted",
  USER_JOINED = "user-joined",
  USER_DISCONNECTED = "user-disconnected",
  SYNC_FILE_STRUCTURE = "sync-file-structure",
  DIRECTORY_CREATED = "directory-created",
  DIRECTORY_UPDATED = "directory-updated",
  DIRECTORY_RENAMED = "directory-renamed",
  DIRECTORY_DELETED = "directory-deleted",
  FILE_CREATED = "file-created",
  FILE_UPDATED = "file-updated",
  FILE_RENAMED = "file-renamed",
  FILE_DELETED = "file-deleted",
  USER_OFFLINE = "offline",
  USER_ONLINE = "online",
  SEND_MESSAGE = "send-message",
  RECEIVE_MESSAGE = "receive-message",
  TYPING_START = "typing-start",
  TYPING_PAUSE = "typing-pause",
  USERNAME_EXISTS = "username-exists",
  REQUEST_DRAWING = "request-drawing",
  SYNC_DRAWING = "sync-drawing",
  DRAWING_UPDATE = "drawing-update",
}

interface User {
  username: string;
  roomId: string;
  status: USER_CONNECTION_STATUS;
  socketId: SocketId;
  cursorPosition: number;
  typing: boolean;
  currentFile: string | null;
}

enum USER_CONNECTION_STATUS {
  ONLINE = "online",
  OFFLINE = "offline",
}

interface SocketContext {
  socket: Socket;
}

// Exporting all the types
export { SocketEvent, SocketContext, SocketId, User, USER_CONNECTION_STATUS };